pyproject.toml
[tool.poetry]
name = "nonebot-pioneer"
version = "0.1.0"
description = "nonebot-pioneer"
authors = ["Autumn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0.a1"
nonebot-adapter-cqhttp = "2.0.0a16"
SQLAlchemy = "^1.4.27"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.0"
nonebot-plugin-test = "^0.3.0"
black = "^21.10b0"
flake8 = "^4.0.1"
flake8-black = "^0.2.3"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/foo.py  # exclude a file named foo.py in the root of the project (in addition to the defaults)
'''

[tool.pytest.ini_options]
minversion = "6.0"
python_files = "*_test.py"
python_classes = "*Test"
addopts = "--cov --cov-report html --cov-report term"

[tool.coverage.run]
source = ["plugins", "tools"]
omit = ["*_test.py"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
